requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
pymongo==4.9.2
pydantic==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
email-validator==2.2.0
